Bug - default tuned profile not idempotent
While tuned can automatically choose and use a self-recommended profile,
it will only do this if there is no active profile already configured.
When it does do this, it will write that chosen profile into the
configuration file and then Puppet and the daemon will fight on the
value stored there.  First Puppet will write a single newline to an
otherwise empty file and then it restarts the service.  The tuned
process chooses a recommended profile and writes the name to the
configuration file and then cycle repeats with the next Puppet agent
run.

To rectify the situation, the profile parameter no longer has a default
so it must be explicitly set with a non-null string value.